requirements.txt
yfinance
openai
requests
pandas
matplotlib
reportlab
pytz
python-dotenv
vaderSentiment

# Web Server Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart

# Additional utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0 